// UraxGUI/URAXWebApp/src/app/model/tax-administration/market-model/market-model.ts
export class marketmodel {
    Id: string;
    mid: string;
    countryName: string;
    countryCode: string;
    specificationMarket: string;
    programMarket: string;
    CurrencyId: string;
    showEdit: boolean;
    showsave: boolean;
    showupdate: boolean;
    cancel: boolean;
    isadd: boolean;
}